Updates the weather cycle and does an initial API call on init
Since OpenWeatherMap (OWM) only updates their data every 10 minutes, we can reduce the update calls and the ability to set cycle timing as a setting. This sets the update cycle to every 5 minutes to allow for the uncertainty of when the data on OWM's side is actually updated.

This also does an initial call for weather data once things are initialized.