Improve BG transitions and add FF support
Changes how background image transitions occur which adds Firefox support (Firefox does not support css transitions for background images as [it's technically not part of the spec](https://www.w3.org/TR/css-transitions-1/#animation-of-property-types-)) and other browsers. This fixes an issue with images scaling to `cover` and an issue with "originalList" not actually being that list (lists got flip-flopped which lead to a single run-through only).